Breakdown tapoo.ts file into multiple easy to manage *.ts files
